refactor(upload): seperate store and context management
